+++ packages/feeds-client/src/feed/feed.ts
@@ -13,6 +13,7 @@ import type {
   ThreadedCommentResponse,
   FollowRequest,
   QueryCommentsRequest,
+  ActivityAddedEvent,
 } from '../gen/models';
 import type { StreamResponse } from '../gen-imports';
 import { StateStore } from '@stream-io/state-store';
@@ -59,6 +60,8 @@ import type {
   PagerResponseWithLoadingStates,
 } from '../types';
 import { checkHasAnotherPage, Constants, uniqueArrayMerge } from '../utils';
+import { handleActivityFeedback } from './event-handlers/activity/handle-activity-feedback';
+import { deepEqual } from '../utils/deep-equal';
 
 export type FeedState = Omit<
   Partial<GetOrCreateFeedResponse & FeedResponse>,
@@ -207,10 +210,15 @@ export class Feed extends FeedApi {
     'user.muted': Feed.noop,
     'user.reactivated': Feed.noop,
     'user.updated': Feed.noop,
+    'feeds.activity.feedback': handleActivityFeedback.bind(this),
   };
 
   protected eventDispatcher: EventDispatcher<WSEvent['type'], WSEvent> =
     new EventDispatcher<WSEvent['type'], WSEvent>();
+  private inProgressGetOrCreate?: {
+    request?: GetOrCreateFeedRequest;
+    promise: Promise<StreamResponse<GetOrCreateFeedResponse>>;
+  };
 
   constructor(
     client: FeedsClient,
@@ -219,6 +227,7 @@ export class Feed extends FeedApi {
     data?: FeedResponse,
     watch = false,
     addNewActivitiesTo: 'start' | 'end' = 'start',
+    public activityAddedEventFilter?: (event: ActivityAddedEvent) => boolean,
   ) {
     super(client, groupId, id);
     this.state = new StateStore<FeedState>({
@@ -266,14 +275,29 @@ export class Feed extends FeedApi {
     return this.indexedActivityIds.has(activityId);
   }
 
+  hasPinnedActivity(activityId: string) {
+    return this.currentState.pinned_activities?.some(
+      (pinnedActivity) => pinnedActivity.activity.id === activityId,
+    );
+  }
+
   async synchronize() {
     const { last_get_or_create_request_config } = this.state.getLatestValue();
     if (last_get_or_create_request_config?.watch) {
+      this.inProgressGetOrCreate = undefined;
       await this.getOrCreate(last_get_or_create_request_config);
     }
   }
 
   async getOrCreate(request?: GetOrCreateFeedRequest) {
+    if (
+      !request?.next &&
+      this.inProgressGetOrCreate &&
+      deepEqual(request, this.inProgressGetOrCreate.request)
+    ) {
+      return this.inProgressGetOrCreate.promise;
+    }
+
     if (this.currentState.is_loading_activities) {
       throw new Error('Only one getOrCreate call is allowed at a time');
     }
@@ -287,7 +311,13 @@ export class Feed extends FeedApi {
     // and pre-populate comments_by_entity_id (once comment_sort and comment_limit are supported)
 
     try {
-      const response = await super.getOrCreate(request);
+      const responsePromise = super.getOrCreate(request);
+
+      if (!request?.next) {
+        this.inProgressGetOrCreate = { request, promise: responsePromise };
+      }
+
+      const response = await responsePromise;
 
       const currentActivityFeeds = [];
       for (const activity of response.activities) {
@@ -296,7 +326,10 @@ export class Feed extends FeedApi {
         }
       }
 
-      this.client.hydrateCapabilitiesCache([response.feed, ...currentActivityFeeds]);
+      this.client.hydrateCapabilitiesCache([
+        response.feed,
+        ...currentActivityFeeds,
+      ]);
 
       if (request?.next) {
         const { activities: currentActivities = [] } = this.currentState;
@@ -383,6 +416,9 @@ export class Feed extends FeedApi {
         is_loading: false,
         is_loading_activities: false,
       });
+      if (!request?.next) {
+        this.inProgressGetOrCreate = undefined;
+      }
     }
   }
 
@@ -841,10 +877,19 @@ export class Feed extends FeedApi {
     return response;
   }
 
+  /**
+   * Fetches the next page of activities for the feed.
+   * @returns The response from the API or `undefined` if there is no next page.
+   */
   async getNextPage() {
     const currentState = this.currentState;
 
+    if (!currentState.next) {
+      return;
+    }
+
     return await this.getOrCreate({
+      ...currentState.last_get_or_create_request_config,
       member_pagination: {
         limit: 0,
       },
@@ -854,8 +899,8 @@ export class Feed extends FeedApi {
       following_pagination: {
         limit: 0,
       },
+      watch: undefined,
       next: currentState.next,
-      limit: currentState.last_get_or_create_request_config?.limit ?? 20,
     });
   }