+++ packages/feeds-client/src/gen/model-decoders/decoders.ts
@@ -83,6 +83,15 @@ decoders.ActivityDeletedEvent = (input?: Record<string, any>) => {
   return decode(typeMappings, input);
 };
 
+decoders.ActivityFeedbackEvent = (input?: Record<string, any>) => {
+  const typeMappings: TypeMapping = {
+    created_at: { type: 'DatetimeType', isSingle: true },
+
+    received_at: { type: 'DatetimeType', isSingle: true },
+  };
+  return decode(typeMappings, input);
+};
+
 decoders.ActivityMarkEvent = (input?: Record<string, any>) => {
   const typeMappings: TypeMapping = {
     created_at: { type: 'DatetimeType', isSingle: true },
@@ -528,7 +537,7 @@ decoders.ChannelMember = (input?: Record<string, any>) => {
 
     pinned_at: { type: 'DatetimeType', isSingle: true },
 
-    user: { type: 'UserResponse', isSingle: true },
+    user: { type: 'User', isSingle: true },
   };
   return decode(typeMappings, input);
 };
@@ -540,6 +549,29 @@ decoders.ChannelMemberLookup = (input?: Record<string, any>) => {
     ban_expires: { type: 'DatetimeType', isSingle: true },
 
     pinned_at: { type: 'DatetimeType', isSingle: true },
+  };
+  return decode(typeMappings, input);
+};
+
+decoders.ChannelMemberResponse = (input?: Record<string, any>) => {
+  const typeMappings: TypeMapping = {
+    created_at: { type: 'DatetimeType', isSingle: true },
+
+    updated_at: { type: 'DatetimeType', isSingle: true },
+
+    archived_at: { type: 'DatetimeType', isSingle: true },
+
+    ban_expires: { type: 'DatetimeType', isSingle: true },
+
+    deleted_at: { type: 'DatetimeType', isSingle: true },
+
+    invite_accepted_at: { type: 'DatetimeType', isSingle: true },
+
+    invite_rejected_at: { type: 'DatetimeType', isSingle: true },
+
+    pinned_at: { type: 'DatetimeType', isSingle: true },
+
+    user: { type: 'UserResponse', isSingle: true },
   };
   return decode(typeMappings, input);
 };
@@ -575,7 +607,7 @@ decoders.ChannelResponse = (input?: Record<string, any>) => {
 
     truncated_at: { type: 'DatetimeType', isSingle: true },
 
-    members: { type: 'ChannelMember', isSingle: false },
+    members: { type: 'ChannelMemberResponse', isSingle: false },
 
     config: { type: 'ChannelConfigWithInfo', isSingle: true },
 
@@ -948,6 +980,23 @@ decoders.FeedResponse = (input?: Record<string, any>) => {
   return decode(typeMappings, input);
 };
 
+decoders.FeedSuggestionResponse = (input?: Record<string, any>) => {
+  const typeMappings: TypeMapping = {
+    created_at: { type: 'DatetimeType', isSingle: true },
+
+    updated_at: { type: 'DatetimeType', isSingle: true },
+
+    created_by: { type: 'UserResponse', isSingle: true },
+
+    deleted_at: { type: 'DatetimeType', isSingle: true },
+
+    own_follows: { type: 'FollowResponse', isSingle: false },
+
+    own_membership: { type: 'FeedMemberResponse', isSingle: true },
+  };
+  return decode(typeMappings, input);
+};
+
 decoders.FeedUpdatedEvent = (input?: Record<string, any>) => {
   const typeMappings: TypeMapping = {
     created_at: { type: 'DatetimeType', isSingle: true },
@@ -1096,7 +1145,7 @@ decoders.GetConfigResponse = (input?: Record<string, any>) => {
 
 decoders.GetFollowSuggestionsResponse = (input?: Record<string, any>) => {
   const typeMappings: TypeMapping = {
-    suggestions: { type: 'FeedResponse', isSingle: false },
+    suggestions: { type: 'FeedSuggestionResponse', isSingle: false },
   };
   return decode(typeMappings, input);
 };
@@ -1240,6 +1289,8 @@ decoders.MessageResponse = (input?: Record<string, any>) => {
     thread_participants: { type: 'UserResponse', isSingle: false },
 
     draft: { type: 'DraftResponse', isSingle: true },
+
+    member: { type: 'ChannelMemberResponse', isSingle: true },
 
     pinned_by: { type: 'UserResponse', isSingle: true },
 
@@ -1707,6 +1758,10 @@ decoders.ReviewQueueItemResponse = (input?: Record<string, any>) => {
     entity_creator: { type: 'EntityCreatorResponse', isSingle: true },
 
     feeds_v2_reaction: { type: 'Reaction', isSingle: true },
+
+    feeds_v3_activity: { type: 'ActivityResponse', isSingle: true },
+
+    feeds_v3_comment: { type: 'CommentResponse', isSingle: true },
 
     message: { type: 'MessageResponse', isSingle: true },